+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix legal vote popover layout issues (!927) 
 - Highlight the protocol icon on first appearance (!934)
 - Fix german Protokoll ausblenden label (!933)
+- Fix audio/video permission denied message (!931)
 - Stay in coffee break view when a timer is started (!915)
 - Disable starting of a legal vote/poll when a coffee break is active (!916)
 - Hide legal vote token for unsubmitted users after vote is finnished. (!926)